aws/sam-yolop/app/requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.8.0+cpu
torchvision==0.9.0+cpu
scipy
tqdm
yacs
Cython
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python-headless>=4.1.2
Pillow
PyYAML>=5.3
scipy>=1.4.1
tensorboardX
seaborn
prefetch_generator
imageio
scikit-learn
